+++ src/controller/controller_user.js
@@ -1,7 +1,7 @@
-const db = require("../../models");
-const Users = db.Users;
-const Op = db.Sequelize.Op;
-const controller = {};
+const db            = require("../../models");
+const Users         = db.Users;
+const Op            = db.Sequelize.Op;
+const controller    = {};
 
 controller.getAll = async (req, res) => {
     const dataUsers = req.query.dataUsers
@@ -11,22 +11,22 @@ controller.getAll = async (req, res) => {
         .then(results => {
             res.send(results)
         })
-    } catch (error) {
+    } catch (err) {
         res.status(500).send({
             message:
-              err.message || "Some error occurred while retrieving tutorials."
+              err.message || "Some error occurred while retrieving users."
           });
     }
 }
 
 controller.register = async (req, res) => {
     const User = {
-        firstName: req.body.firstName,
-        lastName: req.body.lastName,
-        email: req.body.email,
-        pass: req.body.pass,
-        createdAt: new Date(),
-        updateAt: new Date()
+        firstName   : req.body.firstName,
+        lastName    : req.body.lastName,
+        email       : req.body.email,
+        pass        : req.body.pass,
+        createdAt   : new Date(),
+        updateAt    : new Date()
     }
 
     try {
@@ -58,8 +58,8 @@ controller.login = async (req, res) => {
     try {
         await Users.findOne({
             where: {
-                email: email,
-                pass: pass
+                email : email,
+                pass  : pass
             }
         })
         .then(results => {
@@ -96,33 +96,54 @@ controller.getUserById = async (req, res) => {
     }
 }
 
-//  controller.updateName = async (req, res) => {
-//      const id = req.params.id;
-//      const firstName = req.body.firstName;
-//      const lastName = req.body.lastName;
+// controller.updateName = async (req, res) => {
+//     const firstName = req.body.firstName;
+//     const lastName = req.body.lastName;
    
-//     try {
-//         await Users.update({lastName: lastName}, {
-//             where: {
-//                 id: id,
-//                 lastName: lastName
-//             }
+//    try {
+//        await Users.findByPk(req.params.id, {
+//         include: [{
+//             model: Users,
+//             as: "Users"
+//         }],
+//        })
+//        .then(Users => {
+//         if (!Users) {
+//             return res.status(404).send("User Not Found")
+//         }
+//         return Users.update({
+//             firstName: firstName,
+//             lastName: lastName
 //         })
-//         .then(() => res.status(203).send("Updated is successfully"))
-//     } catch (error) {
-//         res.status(404).send("Update is error")
-//     }
-//  }
+//         .then(() => res.status(200).send(Users))
+//         .catch((error) => res.status(400).send(error))
+//        })
+//    } catch (error) {
+//        res.status(404).send("Update is error")
+//    }
+// }
 
 controller.DeleteUser = async (req, res) => {
     const id = req.params.id;
     try {
-        await Users.destroy({
-            where: {
-                id: id
+        await Users.findByPk(id)
+        .then(results => {
+            if(results) {
+                Users.destroy({
+                    where: {
+                        id: id
+                    }
+                })
+                .then((results) => {
+                    res.status(204).send({
+                        delete: results
+                    })
+                })
+            } else {
+                res.send("There's not data")
             }
         })
-        .then(() => res.status(204).send("Delete user's successfully"))
+
     } catch (error) {
         res.status(400).send("error")
     }
